+++ lib/class/stats.class.php
@@ -180,7 +180,7 @@ class Stats
                 return true;
             }
             // if you've recorded in less than 5 seconds i don't believe you
-            if (($row['date'] < $time && $row['date'] > ($time - 5))) {
+            if (($row['date'] < $time && $row['date'] > ($time - 5)) && $row['count_type'] !== 'download') {
                 debug_event('stats.class', 'Too fast! Skipping {' . (string) $object_id . '} date: ' . (string) $time, 5);
 
                 return true;
@@ -326,6 +326,11 @@ class Stats
         $diff      = $date - (int) $previous['date'];
         $item_time = $object->time;
         $skip_time = AmpConfig::get_skip_timer($previous['time']);
+
+        // if your last song is 30 seconds and your skip timer if 40 you don't want to keep skipping it.
+        if ($previous['time'] > 0 && $previous['time'] < $skip_time) {
+            return true;
+        }
 
         // this object was your last play and the length between plays is too short.
         if ($previous['object_id'] == $object->id && $diff < ($item_time)) {